Initial version 0.0.1 commit
Creates an input box to accept a google sheets key, when submitted/used will retrieve data from that sheet (if it's using the template for rotisserie draft) and will strikethrough picks made so far and bold the unpicked cards. Will also use the emojis in names provided in the draft to highlight who made the picks. Currently highlights names without emojis with a '❌' symbol.

Additionally the script stores the search previously made so that on refresh you don't have to go find that link, but can just hit submit again on the form.